Rename the files managing DB entities to "coffeeBeansAPI", "enhancedCoffeeBeansAPI" and "recipesAPI".
